backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
flask-sqlalchemy==3.1.1
flask-migrate==4.0.5
psycopg[binary]==3.1.18
python-dotenv==1.0.0
requests==2.31.0
PyJWT==2.8.0
gunicorn==21.2.0
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-api-python-client==2.111.0

